Change back to float for price

+++ backend/product_service/src/functions/get_product_by_id.py
@@ -75,7 +75,7 @@ def handler(event: Dict[str, Any], context: Any, dynamodb = None) -> Dict[str, A
             'id': product_id,
             'title': product['title'],
             'description': product['description'],
-            'price': int(product['price']),
+            'price': float(product['price']),
             'count': int(stock_count)
         }
